requirements.txt
discord.py
redis
requests
beautifulsoup4
python-dotenv